pyproject.toml
[project]
name = "pixels"
version = "0.1.0"
description = "Real-time collaborative pixel art application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "boto3>=1.34.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    "pillow>=10.1.0",
    "numpy>=1.26.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "fakeredis>=2.20.0",
    "moto>=4.2.0",
    "freezegun>=1.2.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[dependency-groups]
dev = [
    "fakeredis>=2.32.1",
    "freezegun>=1.5.5",
    "httpx>=0.28.1",
    "moto>=5.1.17",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.5",
]
